Refactor constant
